pyproject.toml
[project]
name = "rover-reunion"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.8",
    "django-browser-reload>=1.21.0",
    "django-cotton>=2.4.0",
    "django-extensions>=4.1",
    "django-htmx>=1.26.0",
    "django-import-export>=4.3.14",
    "django-jazzmin>=3.0.1",
    "django-tailwind-cli>=4.4.2",
    "honeypot>=0.1.1",
    "pillow>=12.0.0",
]
